+++ R/plot_log_bootstrap_distributions.R
@@ -12,15 +12,26 @@ plot_log_bootstrap_distributions <- function(dat,
                                              tag,
                                              listFileName,
                                              graph_theme,
-                                             save_dir) {
-    messager(cc,": Saving log bootstrap plot with distributions.")
-    requireNamespace("grDevices")
+                                             save_dir = file.path(
+                                                 tempdir(),
+                                                 paste0("BootstrapPlots",
+                                                        "_for_transcriptome")),
+                                             height=3.5,
+                                             width=3.5) {
     requireNamespace("ggplot2")
     requireNamespace("reshape2")
+    
+    messager(cc,": Saving log bootstrap plot with distributions.")
+    #### Save path ####
+    pdf_path <- file.path(
+        save_dir,
+        sprintf("bootDists_LOG_%s___%s____%s.pdf",
+                tag, listFileName, cc
+        ))
+    dir.create(dirname(pdf_path), showWarnings = FALSE, recursive = TRUE)
     # - First get the ordered gene names
     rownames(dat) <- dat$Gnames
     datOrdered <- data.frame(GSym = rownames(dat), Pos = seq_len(dim(dat)[1]))
-
     # - Arrange the data frame for plotting
     melt_boot <- reshape2::melt(as.matrix(exp_mats[[cc]]))
     colnames(melt_boot) <- c("Rep", "Pos", "Exp")
@@ -28,8 +39,7 @@ plot_log_bootstrap_distributions <- function(dat,
     melt_boot <- merge(melt_boot, datOrdered, by = "Pos")
     melt_boot$GSym <- factor(as.character(melt_boot$GSym),
         levels = as.character(datOrdered$GSym)
-    )
-
+    ) 
     # - Prepare the values of the list genes to be plotted as red dots
     actVals <- data.frame(
         Pos = as.factor(seq_len(length(hit_exp))),
@@ -38,8 +48,7 @@ plot_log_bootstrap_distributions <- function(dat,
     actVals <- merge(actVals, datOrdered, by = "Pos")
     actVals$GSym <- factor(as.character(actVals$GSym),
         levels = as.character(datOrdered$GSym)
-    )
-
+    ) 
     # - Determine whether changes are significant
     p <- rep(1, max(melt_boot$Pos))
     for (i in seq_len(max(melt_boot$Pos))) {
@@ -52,27 +61,21 @@ plot_log_bootstrap_distributions <- function(dat,
     actVals <- cbind(actVals[order(actVals$Pos), ], ast)
     # - Plot the graph!
     wd <- 1 + length(unique(melt_boot[, 4])) * 0.175
-    pdf_path <- file.path(
-        save_dir,
-        sprintf("bootDists_LOG_%s___%s____%s.pdf",
-                 tag, listFileName, cc
-    ))
-    grDevices::pdf(pdf_path, width = wd, height = 4)
-    print(
-        ggplot(melt_boot) +
-            geom_boxplot(aes_string(x = "GSym", y = "Exp"), outlier.size = 0) +
-            graph_theme +
-            theme(axis.text.x = element_text(angle = 90, hjust = 1)) +
-            geom_point(aes_string(x = "GSym", y = "vals"),
-                col = "red",
-                data = actVals
-            ) +
-            geom_text(aes_string(x = "GSym", y = "vals", label = "ast"),
-                colour = "black", col = "black", data = actVals
-            ) +
-            ylab("Expression in cell type (%)\n") +
-            xlab("Least specific --> Most specific") +
-            scale_y_log10(labels = myScalesComma, limits = c(0.01, 100))
-    )
-    grDevices::dev.off()
+    
+    gp <-  ggplot(melt_boot) +
+        geom_boxplot(aes_string(x = "GSym", y = "Exp"), outlier.size = 0) +
+        graph_theme +
+        theme(axis.text.x = element_text(angle = 90, hjust = 1)) +
+        geom_point(aes_string(x = "GSym", y = "vals"),
+                   col = "red",
+                   data = actVals
+        ) +
+        geom_text(aes_string(x = "GSym", y = "vals", label = "ast"),
+                  colour = "black", col = "black", data = actVals
+        ) +
+        ylab("Expression in cell type (%)\n") +
+        xlab("Least specific --> Most specific") +
+        scale_y_log10(labels = myScalesComma, limits = c(0.01, 100))
+    return(list(plot=gp,
+                path=pdf_path))
 }